feat: move struct logic
Some of the logic dealing with the mechanics of the Config and settings structs were moved to the modules where the structs are defined to reduce some of the noise in the other modules. In addition, some struct properties were made private and access to that data is made available through functions in order to hide some complexity from the consumer.